labs/lab5/requirements.txt
numpy==1.13.3
progressbar2==3.34.3
python-utils==2.2.0
scipy==1.0.0
six==1.11.0
